pyproject.toml
[project]
name = "verdandi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.14",
    "fastapi>=0.116.1",
    "fastapi-cli>=0.0.8",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
]

[project.scripts]
docgen = "scripts.docgen:run"

[dependency-groups]
dev = [
    "fastapi[standard]>=0.116.1",
    "ruff>=0.12.4",
    "ty>=0.0.1a15",
]

[tool.uv]
package = true
default-groups = []
